update views_count front

+++ database/migrations/2021_07_27_015403_add_column_view_count_to_products_table.php
@@ -14,7 +14,7 @@ class AddColumnViewCountToProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('views_count')->default(0);
+            $table->float('views_count')->default(0);
         });
     }
